BAU: added new utility class to mirror other integration test repos
